corrected the indentation  focus on understanding

+++ delete.php
@@ -1,15 +1,15 @@
 <?php
 
 require_once 'dbconnect.php';
+if (isset($_GET['id'])) {
+    $id = $_GET['id'];
 
-$id = $_GET['id'];
+    $query = "DELETE FROM contact WHERE id='" . $id . "'";
+    $select = mysqli_query($conn, $query);
 
-$query = "DELETE FROM contact WHERE id='" . $id . "'";
-$select = mysqli_query($conn, $query);
-
-if ($select) {
-    header("Location:display.php");
-} else {
-    echo 'Kindly check your query';
+    if ($select) {
+        header("Location:display.php");
+    } else {
+        echo 'Kindly check your query';
+    }
 }
-
